Made Range Annotation larger

+++ src/net/homelinux/chaoswg/io/makejavauseful/annotations/number/Range.java
@@ -9,6 +9,6 @@ import static java.lang.annotation.RetentionPolicy.*;
 @Target({FIELD, PARAMETER, LOCAL_VARIABLE, METHOD})
 @Retention(RUNTIME)
 public @interface Range {
-  int start();
-  int end ();
+  long start();
+  long end ();
 }
